Service2 sada zahteva da je admin ulogovan

+++ src/main/java/com/raf/airportflightservice/service/impl/AirplaneService.java
@@ -6,6 +6,7 @@ import com.raf.airportflightservice.repository.AirplaneRepository;
 import com.raf.airportflightservice.repository.FlightRepository;
 import com.raf.airportflightservice.service.IAirplaneService;
 import com.raf.airportflightservice.utils.UtilsMethods;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -19,19 +20,31 @@ public class AirplaneService implements IAirplaneService {
         this.airplaneRepository = airplaneRepository;
     }
 
-    @Override
-    public Boolean addAirplane(Airplane airplane) {
-        airplaneRepository.saveAndFlush(airplane);
-        return true;
+    private Boolean checkIfAdmin(String token) {
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", token);
+        System.out.println(headers);
+        ResponseEntity<Object> responseEntity = UtilsMethods.sendGetHeader("http://localhost:8081/is_admin", headers);
+        Boolean isAdmin = (Boolean) responseEntity.getBody();
+        return isAdmin;
     }
 
     @Override
-    public Boolean deleteAirplane(Long id) {
+    public Boolean addAirplane(Airplane airplane, String token) {
+        if(checkIfAdmin(token)) {
+            airplaneRepository.saveAndFlush(airplane);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public Boolean deleteAirplane(Long id, String token) {
         Flight flight = new Flight();
         flight.setAirplaneId(id);
         ResponseEntity<Object> responseEntity = UtilsMethods.sendPost("http://localhost:8082/flight/search", flight);
         List<Flight> flights = (List<Flight>) responseEntity.getBody();
-        if(flights.size() == 0) {
+        if(flights.size() == 0 && checkIfAdmin(token)) {
             airplaneRepository.deleteById(id);
             return true;
         }
